requirements.txt
streamlit==1.40.1
torch==2.4.1
torchvision==0.19.1
opencv-python==4.10.0.84
pygame==2.6.1
Pillow==10.4.0
numpy==1.24.3
requests==2.32.3
